pyproject.toml
[project]
name = "marty"
version = "0.1.0"
description = "Marty, the Dungeon Books RCS Wizard - AI-powered SMS chatbot for book recommendations"
readme = "README.md"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.104.0",
    "hypercorn>=0.16.0", # ASGI server with dual stack IPv4/IPv6 support
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.29.0", # for PostgreSQL async support
    "aiosqlite>=0.19.0", # for SQLite async support
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6", # for form data
    "httpx>=0.25.0", # for HTTP requests to external APIs
    "gql>=3.4.0", # GraphQL client for Hardcover API
    "aiohttp>=3.9.0", # async HTTP for GraphQL transport
    "greenlet>=3.2.3",
    "psycopg2-binary>=2.9.10",
    "anthropic>=0.57.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    "tests",
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    ".ruff_cache",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.12.3",
]
